requirements.txt
# Langchain Core
langchain 
langchain-core 

# OpenAI Integration
langchain-openai 
openai 

# Anthropic Integration 
langchain-anthropic  

# Google Gemini (PaLM) Integration 
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment variable management
python-dotenv

# Machine Learning Utilities 
numpy
scikit-learn   
